+++ backend/fms_core/models/dataset_file.py
@@ -14,7 +14,7 @@ class DatasetFile(TrackedModel):
 
     readset = models.ForeignKey(Readset, on_delete=models.PROTECT, help_text="Readset of the file.", related_name="files")
     file_path = models.CharField(max_length=STANDARD_FILE_PATH_LENGTH, help_text="Path to the dataset file.")
-    size = models.IntegerField(null=False, help_text="Size of the dataset file.")
+    size = models.BigIntegerField(null=False, help_text="Size of the dataset file.")
 
     class Meta:
         constraints = [
